vion-test-senario/requirements.txt
Django==5.1
requests==2.32.3
djangorestframework==3.15.2
drf-yasg==1.21.8
python-dotenv==1.0.1